🔧 refactor: #12 JSdoc 주석 추가
