+++ src/components/icons/business/index.ts
@@ -1,4 +1,4 @@
-export { default as UserIcon } from './CustomerIcon';
 export { default as JourneyIcon } from './JourneyIcon';
 export { default as BookIcon } from './OpenBookIcon';
+export { default as UserIcon } from './UserIcon';
 export { default as VersionIcon } from './VersionIcon';
