+++ src/components/icons/navigation/index.ts
@@ -1,6 +1,6 @@
 export { default as BackIcon } from './BackIcon';
+export { default as ChevronRightIcon } from './ChevronRightIcon';
 export { default as EllipsisIcon } from './EllipsisIcon';
 export { default as LogoutIcon } from './LogOutIcon';
-export { default as ChevronRightIcon } from './MoreThanIcon';
 export { default as SearchIcon } from './SearchIcon';
 export { default as SettingsIcon } from './SettingsIcon';
